그래프 겹치는 현상 모두 해결
