+++ src/component/main/mainGraph.jsx
@@ -4,6 +4,7 @@ import Kospi from "./graphlist/kospi";
 import Kosdaq from "./graphlist/kosdaq";
 import Nasdaq from "./graphlist/nasdaq";
 import Spx from "./graphlist/spx";
+import axios from "axios";
 
 const ChartBox = styled.div`
   display: flex;
@@ -60,6 +61,50 @@ const RightArrow = styled(ArrowButton)`
 const MainGraph = () => {
   const [activeButton, setActiveButton] = useState(1);
 
+  const [lista, setLista] = useState(null);
+  const [audioBuffer, setAudioBuffer] = useState(null); //audio 파일 저장
+  const [isPlaying, setIsPlaying] = useState(false); //그래프 음향 출력 중복 방지
+
+  useEffect(() => {
+    //lista 저장 후 데이터를 소리로 변환
+    if (lista !== null) {
+      axios
+        .post(
+          `https://stalksound.store/sonification/data_to_sound/`,
+          {
+            lista: lista,
+          },
+          { responseType: "arraybuffer" }
+        ) //arraybuffer 형태로 받아서
+        .then(async (res) => {
+          //AudioContext 생성
+          const audioContext = new (window.AudioContext ||
+            window.webkitAudioContext)();
+          const decodedBuffer = await audioContext.decodeAudioData(res.data); //decode
+          setAudioBuffer(decodedBuffer); //디코딩된 정보 저장
+        })
+        .catch((e) => {
+          console.log(e);
+        });
+    }
+  }, [lista]);
+
+  //그래프 음향 출력
+  const playAudio = () => {
+    if (!isPlaying && audioBuffer) {
+      setIsPlaying(true);
+      const audioContext = new (window.AudioContext ||
+        window.webkitAudioContext)();
+      const source = audioContext.createBufferSource();
+      source.buffer = audioBuffer;
+      source.connect(audioContext.destination);
+      source.onended = () => {
+        setIsPlaying(false); //재생 끝날 경우 false로 reset
+      };
+      source.start(0);
+    }
+  };
+
   useEffect(() => {
     const interval = setInterval(() => {
       setActiveButton((prevButton) => (prevButton % 4) + 1);
@@ -78,36 +123,40 @@ const MainGraph = () => {
     setActiveButton((prevButton) => ((prevButton - 2 + 4) % 4) + 1);
   };
 
+  const highFunction = (text) => {
+    setLista(text);
+  };
+
   const renderInfo = () => {
     switch (activeButton) {
       case 1:
         return (
           <>
-            <Kospi />
+            <Kospi propFunction={highFunction} />
           </>
         );
       case 2:
         return (
           <>
-            <Kosdaq />
+            <Kosdaq propFunction={highFunction} />
           </>
         );
       case 3:
         return (
           <>
-            <Spx />
+            <Spx propFunction={highFunction} />
           </>
         );
       case 4:
         return (
           <>
-            <Nasdaq />
+            <Nasdaq propFunction={highFunction} />
           </>
         );
       default:
         return (
           <>
-            <Kospi />
+            <Kospi propFunction={highFunction} />
           </>
         );
     }
@@ -115,7 +164,7 @@ const MainGraph = () => {
 
   return (
     <>
-      <ChartBox>
+      <ChartBox onClick={playAudio}>
         <LeftArrow onClick={handlePreviousChart}>&#8249;</LeftArrow>
         <RightArrow onClick={handleNextChart}>&#8250;</RightArrow>
         {/* <ChartType> 코스피</ChartType> */}
